delete the PuLp solver instances when done, store any solutions as a simple dictionary
put patterns in a separate file
